Update AutoPrompt.cs
Used 'EnumerateFileSystemEntries' instead of separate searches for files and directories. Also used HashSet instead of List for the returned strings. Used HashSet's 'UnionWith', which is equivalent to AddRange.
There isn't much difference in the speed of adding to HashSet, but searching it for something as simple as strings should be MANY times faster that List (especially for large collections).